Replace defines for options with in situ values
The values are now assigned at compiled time. Hence it is not necessary
anymore to check for valid command line values at runtime and assign the
default value if necessary. Now it is sufficient to assign a new value
only if there actually is one.

This is also some preliminary work for other sources of configuration
values.