Grabbing the modifiers keys is not necessary
The event handler grabs the whole keyboard, which means that we receive
key{press,release} events for all keys, even the modifier keys.
Therefore it is not necessary to exclusively grab the modifier keys.
However, we still need to resolve the modifers to keycodes using their
mappings to be able to act accordingly in the handler loop.